moved config to main
