+++ README.md
@@ -1,3 +1,7 @@
 # mwarehouse
 
 Media content storing & management system
+
+Before run set environment variables:
+export REGION="region-name"
+export TEMP_BUCKET_NAME="s3-bucket-name"
