+++ src/service/docker-compose.yaml
@@ -5,6 +5,5 @@ services:
     ports:
       - "8765:8765"
     environment:
-      - TEST_MODE=1
       - REGION="us-east-2"
-      - BUCKET_NAME="foundation-13-temporary"
+      - TEMP_BUCKET_NAME="foundation-13-temporary"
